requirements.txt

pydantic-ai
chromadb
python-dotenv
numpy

pydantic